Added 2 new props to support the force complete functionality.
By using the callback approach, now the swipe gesture can be completed programmatically. This will be helpful to persist the button state a completed operation. When there is a orientation change, the button state will be updated. By using the force complete, we can now show the button as in completed state. Updated the documentation.