Fix animation related issues and improve the swipe performance
Instead of every container width change goes through an animation with a duration, divide into 2 parts. While swipe is in progress, we use 0 duration all the time and for the animations we use the provided or default value i.e. 400ms. This will also fix the swipe thumb frozen in the middle on real devices. (#149)